Refactored elements of HomePage.js | moved utility function from HomePage.js to formatting_utils.js Renamed LandingPage directory to 'HomePage' to correspond with its components.
