+++ src/components/AppBarComponent.js
@@ -4,7 +4,7 @@ import LeftMenuBar from './LeftMenuBar'
 import Grid from '@material-ui/core/Grid'
 import React from 'react'
 import MainPage from './MainPage'
-import { HomePage } from './LandingPage/HomePage'
+import { HomePage } from './HomePage/HomePage'
 import { EraseQuestionnaireButton } from '../navigationBarButtons/EraseQuestionnaireButton'
 import { RenderQuestionnaireButton } from '../navigationBarButtons/RenderQuestionnaireButton'
 import { SaveQuestionnaireButton } from '../navigationBarButtons/SaveQuestionnaireButton'
