+++ src/components/HomePage/HomePage.js
@@ -4,7 +4,6 @@ import {QuestionnaireDetails} from "./QuestionnaireDetails";
 import {QuestionnaireList} from "./QuestionnaireList";
 
 export const HomePage = () => {
-
     const [currentQuestionnaireKey, setCurrentQuestionnaireKey] = useState(null);
     return <Grid data-cy="landingPage" container spacing={10} style={{padding: 100}}>
         <Grid item xs={8}>
@@ -15,14 +14,3 @@ export const HomePage = () => {
         </Grid>
     </Grid>
 };
-
-
-export const alignInGrid = (spacing, ...renderedComponents) => {
-    return <Grid container direction="row" alignItems="center" spacing={spacing}>
-        {renderedComponents.map((component, index) => (
-            <Grid key={index} item>
-                {component}
-            </Grid>))}
-    </Grid>
-};
-
